Check for null before attempting to kill the process
r3480

+++ server-coreless/src/main/java/org/openqa/selenium/server/browserlaunchers/DestroyableRuntimeExecutingBrowserLauncher.java
@@ -40,6 +40,7 @@ public class DestroyableRuntimeExecutingBrowserLauncher extends AbstractBrowserL
 
     /** Kills the process */
     public void close() {
+    	if (process == null) return;
         AsyncExecute.killProcess(process);
     }
